Refactor: move eval_client, simplify critic variants, schema improvements
- Move eval_client.py from props/agents/ to props/agents/critic_dev/
  (only used by critic_dev agents)
- critic.bzl: use PROMPT_TEMPLATE_PATH with runfiles path instead of
  genrule rename; variant prompts are now data deps on :main
- Rename grading_complete (bool) to missing_grading_edges (int count)
  in recall_by_run view for better progress visibility
- Add DB CHECK constraints for non-empty baseline_image_digests and
  allowed_examples on improvement type configs
- Make agent_run_id non-optional in run_agent_loop (always available)
- Update E2E_TESTING.md to reflect testcontainers migration
