+++ props/agents/BUILD.bazel
@@ -3,7 +3,7 @@
 load("@rules_python//python:defs.bzl", "py_library")
 load("//tools/testing:docker.bzl", "docker_py_test")
 
-# NOTE: No aggregator target - use specific per-file targets like :runtime, :eval_client
+# NOTE: No aggregator target - use specific per-file targets like :runtime, :schema
 # This is the Gazelle-compatible pattern (python_generation_mode = file)
 
 py_library(
@@ -67,23 +67,5 @@ docker_py_test(
         "@pypi//pytest",
         "@pypi//pytest_asyncio",  # keep
         "@pypi//pytest_bazel",
-    ],
-)
-
-py_library(
-    name = "eval_client",
-    srcs = ["eval_client.py"],
-    imports = ["../.."],
-    visibility = ["//props:__subpackages__"],
-    deps = [
-        ":runtime",
-        "//props/core:eval_api_models",
-        "//props/core:ids",
-        "//props/core/models:examples",
-        "//props/db:database",
-        "//props/db:examples",
-        "//props/db:models",
-        "@pypi//httpx",
-        "@pypi//sqlalchemy",
     ],
 )
